More code synchronization
-Put PCH inline with main repos
-Add ability to enable and disable AHBot and Playerbots
-Add new build types
-Block GCC 11.2
-Update utf8 lib to latest
-Bump realmd conf version
-Update GitHub actions to remove deprecated actions
-Add missing sql files for unused logs db for install script
-Switch authentication to SRP6
-Sync realm server so it can handle all versions from Vanilla through Cata
-Update extractor scripts
-Update recast navigation files
-Add RecastDemoMod to view map/vmap/mmap files

Credit to Warlockbugs, killerwife, and many others for the changes

+++ src/game/MotionGenerators/MoveMap.h
@@ -23,11 +23,12 @@
 #include <Detour/Include/DetourAlloc.h>
 #include <Detour/Include/DetourNavMesh.h>
 #include <Detour/Include/DetourNavMeshQuery.h>
+#include <mutex>
 
 class Unit;
 
 //  memory management
-inline void* dtCustomAlloc(int size, dtAllocHint /*hint*/)
+inline void* dtCustomAlloc(size_t size, dtAllocHint /*hint*/)
 {
     return (void*)new unsigned char[size];
 }
